load stim parameter values in TIMIT and Tone tokenizers
