+++ nsds_lab_to_nwb/components/stimulus/tokenizers/tone_tokenizer.py
@@ -1,3 +1,4 @@
+from nsds_lab_to_nwb.common.io import read_mat_file
 from nsds_lab_to_nwb.components.stimulus.tokenizers.base_tokenizer import BaseTokenizer
 
 
@@ -13,6 +14,11 @@ class ToneTokenizer(BaseTokenizer):
         self.custom_trial_columns = [('sb', 'Stimulus (s) or baseline (b) period'),
                                      ('frq', 'Stimulus Frequency'),
                                      ('amp', 'Stimulus Amplitude')]
+
+    def _load_stim_parameters(self):
+        stim_params_path = self.stim_configs['stim_params_path']
+        stim_vals = tone_stimulus_values(stim_params_path)
+        return stim_vals
 
     def _tokenize(self, stim_vals, stim_onsets,
                   *, audio_start_time, audio_end_time, rec_end_time):
@@ -72,3 +78,38 @@ class ToneTokenizer(BaseTokenizer):
                                    amp=none_str))
 
         return trial_list
+
+
+def tone_stimulus_values(mat_file_path):
+    """adapted from mars.configs.block_directory
+
+    Parameters
+    ----------
+    mat_file_path : path
+        full path to a .mat file that contains stim_values.
+
+    Returns
+    -------
+    stim_vals : ndarray (n, 2)
+        a 2D array with two columns (NOTE: changed from legacy behavior)
+        `stim_vals[:, 0]` are the amplitudes,
+        `stim_vals[:, 1]` are the frequencies of the tones.
+    """
+    sio = read_mat_file(mat_file_path)
+    stim_vals = sio['stimVls'][:].astype(int)
+
+    # check dimension
+    shape = stim_vals.shape
+    if not (len(shape) == 2) and (2 in shape):
+        # should be a 2D array, with 2 columns or 2 rows
+        raise ValueError('stim_vals dimension mismatch')
+    if shape[0] == 2:
+        # changed from legacy behavior (now 2 columns; was 2 rows in mars)
+        stim_vals = stim_vals.T
+
+    # this offset value comes from mars; what is this?
+    # variable naming (amp_offset) was by JHB and could be wrong
+    amp_offset = 8
+    stim_vals[:, 0] = stim_vals[:, 0] + amp_offset
+
+    return stim_vals
